script-breakdown
These changes were made to add comments and documentation to explain the purpose and function of the code. This was done to help me better understand the code and looking back to track my progress. In the spirit of the principle of "if you can't explain it, you don't understand it".